% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{diff_test}
\alias{diff_test}
\title{Perform comparative analysis using T-Test}
\usage{
diff_test(met, group_factors)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factors}{character vector of group factors}
}
\value{
adds results from comparative analysis to metadata slot
}
\description{
Perform comparative analysis using T-Test
}
\examples{
diff_test(met_example, group_factors = c("tumor_normal","random_gender"), references=c("T","random_male"))
}
