% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_matrix.R
\name{diss_matrix}
\alias{diss_matrix}
\title{Construct dissimilarity matrix for metabolic correlation network}
\usage{
diss_matrix(met)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis}
}
\value{
MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis and dissimilarity matrix
}
\description{
Construct dissimilarity matrix for metabolic correlation network
}
\examples{
diss_matrix(met_example)
}
