% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SMPDBanno.R
\name{get_SMPDBanno}
\alias{get_SMPDBanno}
\title{Annotation using Small Molecule Pathway Database (SMPDB)}
\usage{
get_SMPDBanno(rowData, column_kegg_id, column_hmdb_id, column_chebi_id)
}
\arguments{
\item{metabolite_ids}{vector of HMDB, KEGG or ChEBI identifiers.}
}
\value{
dataframe containing SMPDB annotation for metabolite identifiers.
}
\description{
Annotation using Small Molecule Pathway Database (SMPDB)
}
\examples{
get_SMPDBanno(as.data.frame(rowData(met_example[["raw"]]))[,1:7],column_kegg_id=6,column_hmdb_id=7,column_chebi_id=NA)

}
