% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_modules.R
\name{identify_modules}
\alias{identify_modules}
\title{Identification of metabolic correlation modules}
\usage{
identify_modules(met, min_module_size)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{min_module_size}{minimal modules size (default: 5 metabolites)}
}
\value{
adds objects for metabolic correlation modules to metadata slot
}
\description{
Identification of metabolic correlation modules
}
\examples{
identify_modules(met_example, min_module_size=5)
}
