% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOI_plot.R
\name{MOI_plot}
\alias{MOI_plot}
\title{Module of interest plot}
\usage{
MOI_plot(met, group_factor, MOI)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis}
}
\value{
MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis and dissimilarity matrix
}
\description{
Module of interest plot
}
\examples{
MOI_plot(met_example, group_factor = "tumor_normal", MOI=2)
}
