% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_plot.R
\name{MS_plot}
\alias{MS_plot}
\title{Module significance plot}
\usage{
MS_plot(met, group_factor)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis}
}
\value{
MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis and dissimilarity matrix
}
\description{
Module significance plot
}
\examples{
MS_plot(met_example, group_factor = "tumor_normal")
}
