% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_heatmap.R
\name{na_heatmap}
\alias{na_heatmap}
\title{Heatmap to visualize missing metabolites across the samples}
\usage{
na_heatmap(met, group_factor, label_colors)
}
\arguments{
\item{met}{MultiAssayExperiment object with slot "raw"}

\item{group_factor}{name of column in colData for grouping}

\item{label_colors}{vector of colors for levels of sample_labels}
}
\value{
heatmap visualizing missing metabolites across the samples
}
\description{
Heatmap to visualize missing metabolites across the samples
}
\examples{
na_heatmap(met_example, group_factor="tumor_normal", label_colors=c("darkseagreen","dodgerblue"))
}
