% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_met.R
\name{normalize_met}
\alias{normalize_met}
\title{Normalize metabolomic data by vsn}
\usage{
normalize_met(met)
}
\arguments{
\item{@param}{met MultiAssayExperiment object with slots "raw" and "imputed"}
}
\value{
MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"
}
\description{
Normalize metabolomic data by vsn
}
\examples{
normalize_met(met)
}
