% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_heatmap.R
\name{outlier_heatmap}
\alias{outlier_heatmap}
\title{Heatmap to visualize outliers in the study set}
\usage{
outlier_heatmap(met, group_factor, label_colors, k)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw" and "imputed"}

\item{group_factor}{name of column in colData for grouping}

\item{label_colors}{vector of colors for levels of sample_label}
}
\value{
heatmap to visualize outliers in the study set
}
\description{
Heatmap to visualize outliers in the study set
}
\examples{
outlier_heatmap(met_example, group_factor="tumor_normal", label_colors=c("darkseagreen","dodgerblue"), k=2)
}
