% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{pca plot}
\usage{
pca_plot(met, group_factor, label_colors)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}

\item{label_colors}{vector of colors for levels of sample_label}
}
\value{
pca plot
}
\description{
pca plot
}
\examples{
pca(met_example, group_factor="tumor_normal", label_colors=c("darkseagreen","dodgerblue"))
}
