% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_plot.R
\name{quality_plot}
\alias{quality_plot}
\title{Quality plot of processing steps}
\usage{
quality_plot(met, group_factor, label_colors)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}

\item{label_colors}{vector of colors for levels of sample_label}
}
\value{
Quality plot of processing steps
}
\description{
Quality plot of processing steps
}
\examples{
quality_plot(met_example, group_factor="tumor_normal", label_colors=c("darkseagreen","dodgerblue"))
}
