% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cluster.R
\name{remove_cluster}
\alias{remove_cluster}
\title{Remove cluster from metabolomic data}
\usage{
remove_cluster(met, cluster)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw" and "impute"}

\item{cluster}{numeric vector of cluster to remove}
}
\value{
MultiAssayExperiment object where cluster x is removed from slots "raw" and "impute"
}
\description{
Remove cluster from metabolomic data
}
\examples{
remove_cluster(met_example,cluster=1)
}
