% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_plot.R
\name{tsne_plot}
\alias{tsne_plot}
\title{tSNE plot}
\usage{
tsne_plot(met, group_factor, label_colors)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}

\item{label_colors}{vector of colors for levels of sample_label}
}
\value{
tSNE plot
}
\description{
tSNE plot
}
\examples{
tsne_plot(met_example, group_factor="tumor_normal", label_colors=c("darkseagreen","dodgerblue"))
}
