% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_boxplot.R
\name{variance_boxplot}
\alias{variance_boxplot}
\title{Boxplot comparing the distribution of variance in metabolic measurements across the pathways}
\usage{
variance_boxplot(met, rowAnnotation)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{rowAnnotation}{character name of rowAnnotation to stratify metabolites}
}
\value{
Boxplot comparing the distribution of variance in metabolic measurements across the pathways
}
\description{
Boxplot comparing the distribution of variance in metabolic measurements across the pathways
}
\examples{
variance_boxplot(met_example, rowAnnotation="SMPDB|Pathway.Name")
}
