% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MS.R
\name{calculate_MS}
\alias{calculate_MS}
\title{Calculate module significance}
\usage{
calculate_MS(met, group_factors)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}
}
\value{
adds module significance measures to metadata slot
}
\description{
Calculate module significance
}
\examples{
calculate_MS(met_example, group_factors = c("tumor_normal","random_gender"))
}
