% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_matrix.R
\name{diss_matrix}
\alias{diss_matrix}
\title{Construct dissimilarity matrix for metabolic correlation network}
\usage{
diss_matrix(met)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}
}
\value{
adds dissimilarity matrix to metadata slot
}
\description{
Construct dissimilarity matrix for metabolic correlation network
}
\examples{
diss_matrix(met_example)
}
