% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SMPDBanno.R
\name{get_SMPDBanno}
\alias{get_SMPDBanno}
\title{Annotation using Small Molecule Pathway Database (SMPDB)}
\usage{
get_SMPDBanno(met, column_kegg_id, column_hmdb_id, column_chebi_id)
}
\arguments{
\item{met}{MultiAssayExperiment object created with create_mae}

\item{column_kegg_id}{Column number containing KEGG ids.}

\item{column_chebi_id}{Column number containing ChEBI ids.}

\item{column_hmbd_id}{Column number containing HMDB ids.}
}
\value{
MultiAssayExperiment object with SMPDB annotation
}
\description{
Annotation using Small Molecule Pathway Database (SMPDB)
}
\examples{
get_SMPDBanno(met_example,column_kegg_id=6,column_hmdb_id=7,column_chebi_id=NA)
}
