% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SMPDBanno.R
\name{get_SMPDBanno}
\alias{get_SMPDBanno}
\title{Annotation using Small Molecule Pathway Database (SMPDB)}
\usage{
get_SMPDBanno(metabolite_ids)
}
\arguments{
\item{metabolite_ids}{vector of HMDB, KEGG or ChEBI identifiers.}
}
\value{
dataframe containing SMPDB annotation for metabolite identifiers.
}
\description{
Annotation using Small Molecule Pathway Database (SMPDB)
}
\examples{
get_SMPDBanno(c("HMDB00538", "HMDB00250"))
get_SMPDBanno(c("C00002", "C00020"))
get_SMPDBanno(c("15422", "16027"))
}
