% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_modules.R
\name{identify_modules}
\alias{identify_modules}
\title{Identification of metabolic correlation modules}
\usage{
identify_modules(met, min_module_size)
}
\arguments{
\item{met}{MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis}
}
\value{
MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed" and metadata slot containing the results of the differential analysis and dissimilarity matrix
}
\description{
Identification of metabolic correlation modules
}
\examples{
identify_modules(met_example)
}
