% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOI_plot.R
\name{MOI_plot}
\alias{MOI_plot}
\title{Module of interest plot}
\usage{
MOI_plot(met, group_factor, MOI, label_colors)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}

\item{MOI}{module of interest}
}
\value{
plot exploring the module of interest
}
\description{
Module of interest plot
}
\examples{
MOI_plot(met_example, group_factor = "tumor_normal", MOI=2, label_colors=c("darkseagreen","dodgerblue"))
}
