% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_plot.R
\name{MS_plot}
\alias{MS_plot}
\title{Module significance plot}
\usage{
MS_plot(met, group_factor, p_value_cutoff, fold_change_cutoff,
  p_adjust = FALSE)
}
\arguments{
\item{met}{fully preprocessed MultiAssayExperiment object with slots "raw", "imputed", "norm" and "norm_imputed"}

\item{group_factor}{name of column in colData for grouping}
}
\value{
Module significance plot
}
\description{
Module significance plot
}
\examples{
MS_plot(met_example, group_factor = "tumor_normal",p_value_cutoff=0.05, p_adjust=FALSE)
}
